@@ -24,16 +24,16 @@
     "@babel/preset-react": "^7.12.5",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.0",
     "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/react-hooks": "^7.0.0",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/react-hooks": "^8.0.1",
     "autoprefixer": "^10.0.1",
-    "babel-jest": "^27.0.1",
+    "babel-jest": "^28.1.2",
     "babel-loader": "^8.1.0",
     "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^10.2.4",
+    "copy-webpack-plugin": "^11.0.0",
     "cross-env": "^7.0.2",
     "css-loader": "^6.2.0",
-    "css-minimizer-webpack-plugin": "^3.0.0",
+    "css-minimizer-webpack-plugin": "^4.0.0",
     "dotenv": "^16.0.0",
     "dotenv-webpack": "^7.1.0",
     "eslint": "^8.9.0",
@@ -49,35 +49,35 @@
     "eslint-plugin-testing-library": "^5.0.5",
     "html-webpack-change-assets-extension-plugin": "^1.3.0",
     "html-webpack-plugin": "git+https://github.com/jmknz-icf/html-webpack-plugin.git",
-    "husky": "^7.0.2",
-    "jest": "^27.0.1",
-    "lint-staged": "^12.3.4",
+    "husky": "^8.0.1",
+    "jest": "^28.1.2",
+    "lint-staged": "^13.0.3",
     "mini-css-extract-plugin": "^2.2.1",
     "node-sass": "^7.0.1",
     "postcss": "^8.1.6",
-    "postcss-loader": "^6.1.1",
+    "postcss-loader": "^7.0.0",
     "postcss-preset-env": "^7.4.1",
     "prettier": "^2.1.2",
-    "react-refresh": "^0.11.0",
-    "react-test-renderer": "^17.0.1",
-    "sass-loader": "^12.0.0",
+    "react-refresh": "^0.14.0",
+    "react-test-renderer": "^18.2.0",
+    "sass-loader": "^13.0.2",
     "style-loader": "^3.2.1",
     "stylelint": "^14.5.2",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-recommended-scss": "^5.0.2",
-    "stylelint-config-standard": "^25.0.0",
+    "stylelint-config-recommended-scss": "^6.0.0",
+    "stylelint-config-standard": "^26.0.0",
     "stylelint-scss": "^4.1.0",
     "webpack": "^5.4.0",
     "webpack-bundle-analyzer": "^4.1.0",
     "webpack-cli": "^4.2.0",
     "webpack-dev-server": "^4.1.0",
-    "webpack-manifest-plugin": "^4.0.2",
+    "webpack-manifest-plugin": "^5.0.0",
     "webpack-merge": "^5.3.0"
   },
   "dependencies": {
-    "axios": "^0.26.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "axios": "^0.27.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.2.1",
     "tailwindcss": "^3.0.23"
   },
